Updated GUI and Game classes with the starts of a basic GUI.
Currently shows the map (grass and path), monsters, and towers. Allows towers to be built by clicking on the map.

Work in progress, only being pushed to demo to other group members.

Also fixed errors and made minor updates to a few other classes in order to fit the design of the GUI and Game better.

Added resources folder and several images taken from http://opengameart.org/ that are being used for the GUI.